Add files via upload
This application creates 2 tabs 1 for graph and other for table the table displays all the content of the map.csv and we can filter it on basis of date and time (We have 2 columns one for date and other for time). We can also export our selected range as csv or pdf for the table. Graph has a navigation toolbar functionality which can we used to zoom in the graph. We can also select which legend we do not want to display and it will toggle off. We can export graph as pdf or PNG. I have also uploaded the csv which I have used